+++ app/Facades/Experiment.php
@@ -4,6 +4,10 @@ namespace App\Facades;
 
 use Illuminate\Support\Facades\Facade;
 
+/*
+ * A wrapper around session that helps with determining whether
+ * to serve an 'experimental' feature.
+ */
 class Experiment extends Facade
 {
     public static function is($name)