+++ resources/lang/sk/user-collection.php
@@ -2,7 +2,8 @@
 
 return array(
     'title'            => 'Obľúbené diela',
-    'empty'            => 'Momentálne nemáte uložené žiadne diela.',
+    'empty'            => 'Momentálne nemáš uložené žiadne diela.',
+    'empty_hint'       => 'Dielo označíš ako obľúbené pomocou hviezdičky',
     'content-intro'    => 'Obľúbené diela sú uložené iba v tomto internetovom prehliadači.',
     'content-usage'    => 'Dielo zo zoznamu odstrániš kliknutím na hviezdičku. Odstránené diela zmiznú po opätovnom načítaní stránky.',
     'share-info'       => 'Z aktuálnych obľúbených diel môžeš vytvoriť tematický výber.',