+++ resources/views/frontend/user-collection/show.blade.php
@@ -18,18 +18,19 @@
                 @if($items->isNotEmpty())
                     <p>{{ trans('user-collection.content-intro') }}</p>
                     <p>{{ trans('user-collection.content-usage') }}</p>
-                    @if (Experiment::is('WEBUMENIA-1654-shared-user-collections-with-scena'))
-                        <p>{{ trans('user-collection.share-info') }}</p>
-                    @endif
+                    <p>{{ trans('user-collection.share-info') }}</p>
                 @else
-                    <p class="alert alert-info text-center">{{ trans('user-collection.empty') }}</p>
+                    <div class="text-center text-muted">
+                        <p>{{ trans('user-collection.empty') }}</p>
+                        <img src="{{ asset('/images/no-image/diela/no-image-o.jpg') }}" class="w-50" />
+                        <p>{{ trans('user-collection.empty_hint') }}&nbsp;<i class="fa fa-star-o"></i>.</p>
+                    </div>
                 @endif
             </div>
         </div>
     </div>
 </div>
 
-@if (Experiment::is('WEBUMENIA-1654-shared-user-collections-with-scena'))
 <user-collections-store v-slot="store">
     <div :class="{ 'bg-gray-300': store.sharedCollections.present }" class="mt-5 py-5">
         <div class="container">
@@ -52,7 +53,7 @@
                 </div>
             </template>
 
-            <div class="text-center mt-5">
+            <div v-if="store.items.present" v-cloak class="text-center mt-5">
                 <a class="btn btn-dark font-light p-3 px-5" href="{{ route('frontend.shared-user-collections.create', ['ids' => request()->ids ] ) }}">
                     {{ trans('user-collection.share') }}
                 </a>
@@ -65,7 +66,6 @@
         </div>
     </div>
 </user-collections-store>
-@endif
 
 <div class="container">
     @unless($items->isEmpty())